requirements.txt
requests
llama-cpp-python
Pillow
playwright
fastapi
uvicorn[standard]
python-multipart
playwright-mcp
mcp-searxng
mcp-text-editor
FileScopeMCP
mcp-language-server
mcp-code-runner
chroma-mcp
tree-sitter
sentence-transformers
tree-sitter-languages
vosk
torch
silero
pyaudio
sounddevice